4.21*: blogilistan laajennus, step10

+++ blogilista/controllers/blogs.js
@@ -46,6 +46,24 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({
+        error: 'token missing or invalid'
+      })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(401).json({
+        error: 'blog already removed'
+      })
+    }
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({
+        error: 'only blogs creator can remove it'
+      })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
